+++ internal/wrappers/scans-mock.go
@@ -17,7 +17,7 @@ func (m *ScansMockWrapper) Create(model *scansRESTApi.Scan) (*scansRESTApi.ScanR
 	}, nil, nil
 }
 
-func (m *ScansMockWrapper) Get(limit, offset uint) (*scansRESTApi.SlicedScansResponseModel, *scansRESTApi.ErrorModel, error) {
+func (m *ScansMockWrapper) Get(limit, offset uint64) (*scansRESTApi.SlicedScansResponseModel, *scansRESTApi.ErrorModel, error) {
 	fmt.Println("Called Get in ScansMockWrapper")
 	return &scansRESTApi.SlicedScansResponseModel{
 		Scans: []scansRESTApi.ScanResponseModel{
